+++ presto-hive/src/main/java/io/prestosql/plugin/hive/HdfsConfigurationInitializer.java
@@ -50,7 +50,7 @@ import static org.apache.hadoop.hdfs.DFSConfigKeys.DFS_DOMAIN_SOCKET_PATH_KEY;
 import static org.apache.hadoop.hive.conf.HiveConf.ConfVars.COMPRESSRESULT;
 import static org.apache.hadoop.io.SequenceFile.CompressionType.BLOCK;
 
-public class HdfsConfigurationUpdater
+public class HdfsConfigurationInitializer
 {
     private final HostAndPort socksProxy;
     private final Duration ipcPingInterval;
@@ -66,13 +66,13 @@ public class HdfsConfigurationUpdater
     private int textMaxLineLength;
 
     @VisibleForTesting
-    public HdfsConfigurationUpdater(HiveClientConfig config)
+    public HdfsConfigurationInitializer(HiveClientConfig config)
     {
         this(config, ignored -> {});
     }
 
     @Inject
-    public HdfsConfigurationUpdater(HiveClientConfig config, S3ConfigurationUpdater s3ConfigurationUpdater)
+    public HdfsConfigurationInitializer(HiveClientConfig config, S3ConfigurationUpdater s3ConfigurationUpdater)
     {
         requireNonNull(config, "config is null");
         checkArgument(config.getDfsTimeout().toMillis() >= 1, "dfsTimeout must be at least 1 ms");
@@ -106,7 +106,7 @@ public class HdfsConfigurationUpdater
         return result;
     }
 
-    public void updateConfiguration(Configuration config)
+    public void initializeConfiguration(Configuration config)
     {
         copy(resourcesConfiguration, config);
