+++ presto-hive/src/main/java/io/prestosql/plugin/hive/authentication/KerberosHadoopAuthentication.java
@@ -13,7 +13,7 @@
  */
 package io.prestosql.plugin.hive.authentication;
 
-import io.prestosql.plugin.hive.HdfsConfigurationUpdater;
+import io.prestosql.plugin.hive.HdfsConfigurationInitializer;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.security.UserGroupInformation;
 
@@ -28,10 +28,10 @@ public class KerberosHadoopAuthentication
 {
     private final KerberosAuthentication kerberosAuthentication;
 
-    public static KerberosHadoopAuthentication createKerberosHadoopAuthentication(KerberosAuthentication kerberosAuthentication, HdfsConfigurationUpdater updater)
+    public static KerberosHadoopAuthentication createKerberosHadoopAuthentication(KerberosAuthentication kerberosAuthentication, HdfsConfigurationInitializer initializer)
     {
         Configuration configuration = getInitialConfiguration();
-        updater.updateConfiguration(configuration);
+        initializer.initializeConfiguration(configuration);
 
         // In order to enable KERBEROS authentication method for HDFS
         // UserGroupInformation.authenticationMethod static field must be set to KERBEROS
