+++ presto-hive/src/test/java/io/prestosql/plugin/hive/HiveTestUtils.java
@@ -121,7 +121,7 @@ public final class HiveTestUtils
 
     public static HdfsEnvironment createTestHdfsEnvironment(HiveClientConfig config)
     {
-        HdfsConfiguration hdfsConfig = new HiveHdfsConfiguration(new HdfsConfigurationUpdater(config, new PrestoS3ConfigurationUpdater(new HiveS3Config())), ImmutableSet.of());
+        HdfsConfiguration hdfsConfig = new HiveHdfsConfiguration(new HdfsConfigurationInitializer(config, new PrestoS3ConfigurationUpdater(new HiveS3Config())), ImmutableSet.of());
         return new HdfsEnvironment(hdfsConfig, config, new NoHdfsAuthentication());
     }
